fix: distribute code to separate files
