+++ simulation_tests.py
@@ -1,0 +1,6 @@
+import unittest
+import qmodel
+
+class TestQmodel(unittest.TestCase):
+    def setUp():
+        self.system = QSystem()
